Add VoterTrait to avoid duplication
